backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
alembic==1.13.1
asyncpg==0.29.0
psycopg==3.1.10
psycopg-binary==3.2.2
SQLAlchemy==2.0.25 
pydantic==2.6.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyJWT==2.8.0